colocate util function close to where they are needed
